requirements.txt
activate
annotated-types
anyio
asgiref
certifi
charset-normalizer
colorama
contourpy
crispy-bootstrap5
cycler
diskcache
distlib
distro
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-environ
django-filter
django-pwdprotect
django-tokenauth
djangorestframework
docker
filelock
FLAML
fonttools
h11
httpcore
httpx
idna
kiwisolver
Markdown
matplotlib
migrations
MouseInfo
mpmath
numpy
openai
packaging
Pillow
platformdirs
psycopg2-binary
pyautogen
PyAutoGUI
pydantic
pydantic_core
PyGetWindow
Pygments
PyMsgBox
pyparsing
pyperclip
PyRect
PyScreeze
python-dateutil
python-dotenv
pytweening
pytz
pywin32
regex
requests
scipy
six
sniffio
sqlparse
sympy
termcolor
tiktoken
tokenauth
tqdm
typing_extensions
tzdata
urllib3
virtualenv
virtualenvwrapper-win